refactor: 디렉토리 합의된 컨벤션에 맞게 수정
