+++ src/pages/PostPage/PostEditPage.tsx
@@ -1,4 +1,4 @@
-import Posts from "../../components/post/Posts";
+import Posts from "./components/Posts";
 
 function PostEditPage() {
