+++ src/pages/PostPage/components/LazyLoading.tsx
