+++ src/pages/PostPage/components/MessageModal.tsx
@@ -1,6 +1,6 @@
 import React, { SetStateAction } from "react";
-import { MessageRetrieve } from "../../DTO/message/MessageRetrieve";
-import formatComparedTime from "../../pages/PostPage/utils/formatComparedTime";
+import { MessageRetrieve } from "../../../DTO/message/MessageRetrieve";
+import formatComparedTime from "../utils/formatComparedTime";
 import SenderInfo from "./SenderInfo";
 
 type props = {
