+++ src/pages/PostPage/components/PlusCard.tsx
@@ -1,4 +1,4 @@
-import plusIcon from "../../assets/icons/plus.svg";
+import plusIcon from "../../../assets/icons/plus.svg";
 
 function PlusCard() {
